+++ test/mocha/12-keystores-find-api.js
@@ -1,9 +1,7 @@
-/*
- * Copyright (c) 2019-2021 Digital Bazaar, Inc. All rights reserved.
+/*!
+ * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
-
-const {keystores} = require('bedrock-kms');
+import {keystores} from '@bedrock/kms';
 
 describe('keystores APIs', () => {
   const mockConfigAlpha = {